Rename namespace to User
