+++ src/Umbraco.Cms.Api.Management/Controllers/User/ByKeyUsersController.cs
@@ -6,7 +6,7 @@ using Umbraco.Cms.Core.Models.Membership;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class ByKeyUserController : UserControllerBase
 {
