+++ src/Umbraco.Cms.Api.Management/Controllers/User/ChangePasswordUsersController.cs
@@ -8,7 +8,7 @@ using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class ChangePasswordUserController : UserControllerBase
 {
