+++ src/Umbraco.Cms.Api.Management/Controllers/User/CreateUsersController.cs
@@ -10,7 +10,7 @@ using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class CreateUserController : UserControllerBase
 {
