+++ src/Umbraco.Cms.Api.Management/Controllers/User/DeleteUsersController.cs
@@ -2,7 +2,7 @@
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class DeleteUserController : UserControllerBase
 {
