+++ src/Umbraco.Cms.Api.Management/Controllers/User/GetAllUsersController.cs
@@ -11,7 +11,7 @@ using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 using Umbraco.New.Cms.Core.Models;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class GetAllUserController : UserControllerBase
 {
