+++ src/Umbraco.Cms.Api.Management/Controllers/User/InviteUsersController.cs
@@ -9,7 +9,7 @@ using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class InviteUserController : UserControllerBase
 {
