+++ src/Umbraco.Cms.Api.Management/Controllers/User/UpdateUserGroupsUsersController.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Api.Management.ViewModels.Users;
 using Umbraco.Cms.Core.Services;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Users;
+namespace Umbraco.Cms.Api.Management.Controllers.User;
 
 public class UpdateUserGroupsUserController : UserControllerBase
 {
